+++ src/Provider/Nominatim/Tests/NominatimTest.php
@@ -26,39 +26,33 @@ class NominatimTest extends BaseTestCase
         return __DIR__.'/.cached_responses';
     }
 
+    /**
+     * @expectedException \Geocoder\Exception\UnsupportedOperation
+     * @expectedExceptionMessage The Nominatim provider does not support IP addresses.
+     */
     public function testGeocodeWithLocalhostIPv4()
     {
-        $provider = Nominatim::withOpenStreetMapServer($this->getMockedHttpClient());
-        $results = $provider->geocodeQuery(GeocodeQuery::create('127.0.0.1'));
-
-        $this->assertInstanceOf('Geocoder\Model\AddressCollection', $results);
-        $this->assertCount(1, $results);
-
-        /** @var Location $result */
-        $result = $results->first();
-        $this->assertInstanceOf('\Geocoder\Model\Address', $result);
-        $this->assertEquals('localhost', $result->getLocality());
-        $this->assertEmpty($result->getAdminLevels());
-        $this->assertEquals('localhost', $result->getCountry()->getName());
+        $provider = Nominatim::withOpenStreetMapServer($this->getMockedHttpClient(), 'Geocoder PHP/Nominatim Provider/Nominatim Test');
+        $provider->geocodeQuery(GeocodeQuery::create('127.0.0.1'));
     }
 
     /**
      * @expectedException \Geocoder\Exception\UnsupportedOperation
-     * @expectedExceptionMessage The Nominatim provider does not support IPv6 addresses.
+     * @expectedExceptionMessage The Nominatim provider does not support IP addresses.
      */
     public function testGeocodeWithLocalhostIPv6()
     {
-        $provider = Nominatim::withOpenStreetMapServer($this->getMockedHttpClient());
+        $provider = Nominatim::withOpenStreetMapServer($this->getMockedHttpClient(), 'Geocoder PHP/Nominatim Provider/Nominatim Test');
         $provider->geocodeQuery(GeocodeQuery::create('::1'));
     }
 
     /**
      * @expectedException \Geocoder\Exception\UnsupportedOperation
-     * @expectedExceptionMessage The Nominatim provider does not support IPv6 addresses.
+     * @expectedExceptionMessage The Nominatim provider does not support IP addresses.
      */
     public function testGeocodeWithRealIPv6()
     {
-        $provider = Nominatim::withOpenStreetMapServer($this->getHttpClient());
+        $provider = Nominatim::withOpenStreetMapServer($this->getHttpClient(), 'Geocoder PHP/Nominatim Provider/Nominatim Test');
         $provider->geocodeQuery(GeocodeQuery::create('::ffff:88.188.221.14'));
     }
 
@@ -67,7 +61,7 @@ class NominatimTest extends BaseTestCase
      */
     public function testGeocodeWithAddressGetsEmptyContent()
     {
-        $provider = Nominatim::withOpenStreetMapServer($this->getMockedHttpClient('<foo></foo>'));
+        $provider = Nominatim::withOpenStreetMapServer($this->getMockedHttpClient('<foo></foo>'), 'Geocoder PHP/Nominatim Provider/Nominatim Test');
         $provider->geocodeQuery(GeocodeQuery::create('Läntinen Pitkäkatu 35, Turku'));
     }
 
@@ -79,7 +73,7 @@ class NominatimTest extends BaseTestCase
         $emptyXML = <<<'XML'
 <?xml version="1.0" encoding="utf-8"?><searchresults_empty></searchresults_empty>
 XML;
-        $provider = Nominatim::withOpenStreetMapServer($this->getMockedHttpClient($emptyXML));
+        $provider = Nominatim::withOpenStreetMapServer($this->getMockedHttpClient($emptyXML), 'Geocoder PHP/Nominatim Provider/Nominatim Test');
         $provider->geocodeQuery(GeocodeQuery::create('Läntinen Pitkäkatu 35, Turku'));
     }
 
@@ -91,7 +85,7 @@ XML;
     <error>Unable to geocode</error>
 </reversegeocode>
 XML;
-        $provider = Nominatim::withOpenStreetMapServer($this->getMockedHttpClient($errorXml));
+        $provider = Nominatim::withOpenStreetMapServer($this->getMockedHttpClient($errorXml), 'Geocoder PHP/Nominatim Provider/Nominatim Test');
         $result = $provider->reverseQuery(ReverseQuery::fromCoordinates(-80.000000, -170.000000));
 
         $this->assertInstanceOf(Collection::class, $result);
@@ -100,7 +94,7 @@ XML;
 
     public function testGetNodeStreetName()
     {
-        $provider = Nominatim::withOpenStreetMapServer($this->getHttpClient());
+        $provider = Nominatim::withOpenStreetMapServer($this->getHttpClient(), 'Geocoder PHP/Nominatim Provider/Nominatim Test');
         $results = $provider->reverseQuery(ReverseQuery::fromCoordinates(48.86, 2.35));
 
         $this->assertInstanceOf('Geocoder\Model\AddressCollection', $results);
@@ -114,7 +108,7 @@ XML;
 
     public function testGeocodeWithRealAddress()
     {
-        $provider = Nominatim::withOpenStreetMapServer($this->getHttpClient());
+        $provider = Nominatim::withOpenStreetMapServer($this->getHttpClient(), 'Geocoder PHP/Nominatim Provider/Nominatim Test');
         $results = $provider->geocodeQuery(GeocodeQuery::create('35 avenue jean de bologne 1020 bruxelles'));
 
         $this->assertInstanceOf('Geocoder\Model\AddressCollection', $results);